app/db/requirements.txt
alembic==1.13.1
SQLAlchemy==2.0.28
python-dotenv==1.0.1
psycopg2==2.9.9
PyMovieDb==0.0.9
requests_html==0.10.0
psycopg==3.1.18
lxml_html_clean